+++ packages/curd-table/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @naive-potato-ui/curd-table
 
+## 1.0.11
+
+### Patch Changes
+
+- config is selectable
+
 ## 1.0.10
 
 ### Patch Changes